refactor: separate source context lifecycle and validation concerns
