+++ .agents/plugins/opencode-aidevops/source-access-runtime.mjs
@@ -9,6 +9,7 @@ import { promisify } from "node:util";
 import { sourceGitArguments, sourceGitEnvironment } from "./source-access-git.mjs";
 import {
   createSourceContextResponder,
+  isUnprivilegedSourceRuntime,
   listenSourceContext,
   prepareSourceContextDirectory,
   SOURCE_CONTEXT_QUERY,
@@ -65,115 +66,130 @@ async function abortable(promise, signal) {
   }
 }
 
-/** V1 plugin SDK adapter: fresh lookup, not cached events or caller-written state. */
-export function createSourceAccessRuntime({ client, directory, scriptsDir, tempDir, enabled = true }) {
-  let endpoint;
-  let starting;
-  let closed = false;
-  let pendingLookups = 0;
-  const respond = createSourceContextResponder({
-    async lookupSession(id, signal) {
-      // The /v1 plugin uses SDK get({path:{id}, query:{directory}}). V2 is
-      // deliberately unsupported; malformed/error responses never prove life.
-      if (closed || pendingLookups >= 8) throw new Error("source context lookup unavailable");
-      pendingLookups++;
-      const lookup = Promise.resolve().then(() =>
-        client.session.get({ path: { id }, query: { directory }, signal }))
-        .finally(() => { pendingLookups--; });
-      // A defective SDK that ignores cancellation cannot hang a tool or free
-      // capacity for unlimited outstanding calls: retain its slot until settled.
-      const result = await abortable(lookup, signal);
-      if (result?.error || !result?.data) throw new Error("source context session is unavailable");
-      return result.data;
-    },
-    sameRepository: (sessionDirectory, root, signal) =>
-      sameRepository(directory, sessionDirectory, root, signal),
-    async verifyOwner(root, sessionId, signal) {
-      const result = await command(join(scriptsDir, "worktree-helper.sh"),
-        ["registry", "verify-owner", root, sessionId], signal);
-      return result.trim() === "VERIFIED";
-    },
-  });
+class SourceAccessRuntime {
+  #config;
+  #respond;
+  #endpoint;
+  #starting;
+  #closed = false;
+  #pendingLookups = 0;
+  #onExit = () => this.close();
 
-  async function start() {
-    if (closed || !enabled) return undefined;
-    if (endpoint) return endpoint;
-    starting ??= (async () => {
-      const listener = await listenSourceContext({
-        directory: prepareSourceContextDirectory(tempDir), respond,
-      });
-      if (closed) {
-        listener.close();
-        return undefined;
-      }
-      endpoint = listener;
-      process.once("exit", close);
-      return listener;
-    })();
+  constructor({ enabled = true, ...config }) {
+    this.#config = { ...config, enabled: Boolean(enabled && isUnprivilegedSourceRuntime()) };
+    this.#respond = createSourceContextResponder({
+      lookupSession: (id, signal) => this.#lookupSession(id, signal),
+      sameRepository: (sessionDirectory, root, signal) =>
+        sameRepository(this.#config.directory, sessionDirectory, root, signal),
+      verifyOwner: (root, sessionId, signal) => this.#verifyOwner(root, sessionId, signal),
+    });
+  }
+
+  async #lookupSession(id, signal) {
+    // The /v1 plugin uses SDK get({path:{id}, query:{directory}}). V2 is
+    // deliberately unsupported; malformed/error responses never prove life.
+    if (this.#closed || this.#pendingLookups >= 8) throw new Error("source context lookup unavailable");
+    this.#pendingLookups++;
+    const { client, directory } = this.#config;
+    const lookup = Promise.resolve().then(() => client.session.get({ path: { id }, query: { directory }, signal }))
+      .finally(() => { this.#pendingLookups--; });
+    // Keep a defective SDK's slot occupied until it actually settles, even
+    // when cancellation releases the caller; outstanding work stays bounded.
+    const result = await abortable(lookup, signal);
+    if (result?.error || !result?.data) throw new Error("source context session is unavailable");
+    return result.data;
+  }
+
+  async #verifyOwner(root, sessionId, signal) {
+    const result = await command(join(this.#config.scriptsDir, "worktree-helper.sh"),
+      ["registry", "verify-owner", root, sessionId], signal);
+    return result.trim() === "VERIFIED";
+  }
+
+  async #start() {
+    if (this.#closed || !this.#config.enabled) return undefined;
+    if (this.#endpoint) return this.#endpoint;
+    this.#starting ??= this.#bindListener();
     try {
-      return await starting;
+      return await this.#starting;
     } finally {
-      starting = undefined;
+      this.#starting = undefined;
     }
   }
 
-  function close() {
-    closed = true;
-    endpoint?.close();
-    endpoint = undefined;
-    process.removeListener("exit", close);
+  async #bindListener() {
+    const listener = await listenSourceContext({
+      directory: prepareSourceContextDirectory(this.#config.tempDir), respond: this.#respond,
+    });
+    if (this.#closed) {
+      listener.close();
+      return undefined;
+    }
+    this.#endpoint = listener;
+    process.once("exit", this.#onExit);
+    return listener;
   }
 
-  async function resolveContext(sessionId, repoRoot, signal = AbortSignal.timeout(4000)) {
-    if (closed || !enabled) return undefined;
+  close() {
+    this.#closed = true;
+    this.#endpoint?.close();
+    this.#endpoint = undefined;
+    process.removeListener("exit", this.#onExit);
+  }
+
+  async resolve(sessionId, repoRoot, signal = AbortSignal.timeout(4000)) {
+    if (this.#closed || !this.#config.enabled) return undefined;
     try {
-      const reply = await respond({ schema: SOURCE_CONTEXT_QUERY, nonce: randomBytes(32).toString("hex"),
+      const reply = await this.#respond({ schema: SOURCE_CONTEXT_QUERY, nonce: randomBytes(32).toString("hex"),
         session_id: sessionId, repo_root: repoRoot }, signal);
-      return closed || signal.aborted ? undefined : reply;
+      return this.#closed || signal.aborted ? undefined : reply;
     } catch {
       return undefined;
     }
   }
 
-  return {
-    // Shell hints locate the channel; they NEVER supply consumer authority.
-    async environment(sessionId) {
-      const env = { AIDEVOPS_SOURCE_CONTEXT_SOCKET: "", AIDEVOPS_SOURCE_CONTEXT_INSTANCE: "" };
-      if (!SESSION_ID.test(sessionId)) return env;
+  // Shell hints locate the channel; they NEVER supply consumer authority.
+  async environment(sessionId) {
+    const env = { AIDEVOPS_SOURCE_CONTEXT_SOCKET: "", AIDEVOPS_SOURCE_CONTEXT_INSTANCE: "" };
+    if (!SESSION_ID.test(sessionId)) return env;
+    try {
+      const listener = await this.#start();
+      if (listener) {
+        env.AIDEVOPS_SOURCE_CONTEXT_SOCKET = listener.socketPath;
+        env.AIDEVOPS_SOURCE_CONTEXT_INSTANCE = sourceContextInstanceId;
+      }
+    } catch {
+      // Unsupported/private-directory failures leave only legacy approval usable.
+    }
+    return env;
+  }
+
+  async contexts(sessionId, paths) {
+    const contexts = new Map();
+    if (this.#closed || !this.#config.enabled || !Array.isArray(paths) || paths.length > 32) return contexts;
+    const signal = AbortSignal.timeout(4000);
+    const roots = new Map();
+    for (const path of new Set(paths)) {
+      if (signal.aborted) break;
       try {
-        const listener = await start();
-        if (listener) {
-          env.AIDEVOPS_SOURCE_CONTEXT_SOCKET = listener.socketPath;
-          env.AIDEVOPS_SOURCE_CONTEXT_INSTANCE = sourceContextInstanceId;
-        }
+        if (typeof path !== "string" || !isAbsolute(path)) continue;
+        const root = await gitPath(dirname(path), "--show-toplevel", signal);
+        if (!roots.has(root)) roots.set(root, await this.resolve(sessionId, root, signal));
+        contexts.set(resolve(path), roots.get(root));
       } catch {
-        // Unsupported/private-directory failures leave only legacy approval usable.
+        // No context for an unavailable path; do not borrow another scope.
       }
-      return env;
-    },
-    resolve: resolveContext,
-    async contexts(sessionId, paths) {
-      const contexts = new Map();
-      if (closed || !enabled || !Array.isArray(paths) || paths.length > 32) return contexts;
-      const signal = AbortSignal.timeout(4000);
-      const roots = new Map();
-      for (const path of new Set(paths)) {
-        if (signal.aborted) break;
-        try {
-          if (typeof path !== "string" || !isAbsolute(path)) continue;
-          const root = await gitPath(dirname(path), "--show-toplevel", signal);
-          if (!roots.has(root)) roots.set(root, await resolveContext(sessionId, root, signal));
-          contexts.set(resolve(path), roots.get(root));
-        } catch {
-          // No context for an unavailable path; do not borrow another scope.
-        }
-      }
-      return signal.aborted || closed ? new Map() : contexts;
-    },
-    handleEvent(input) {
-      const event = input?.event;
-      if (event?.type === "server.instance.disposed" && event.properties?.directory === directory) close();
-    },
-    close,
-  };
+    }
+    return signal.aborted || this.#closed ? new Map() : contexts;
+  }
+
+  handleEvent(input) {
+    const event = input?.event;
+    if (event?.type === "server.instance.disposed" && event.properties?.directory === this.#config.directory) this.close();
+  }
+}
+
+/** V1 plugin SDK adapter: fresh lookup, not cached events or caller-written state. */
+export function createSourceAccessRuntime(config) {
+  return new SourceAccessRuntime(config);
 }
